requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
google==2.0.3
google-api-core==1.16.0
google-auth==1.12.0
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
mysqlclient==1.4.6
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
